renamed rule name, now throw error on every bad instance
