+++ lib/options_manifest.json
@@ -39,7 +39,7 @@
         "desc": "If set, the doctype element must specify html5.",
         "default": true
     },
-    "head-content-model": {
+    "head-valid-content-model": {
         "desc": "If set, the content-model of the HEAD element will be enforced: only BASE, LINK, META, NOSCRIPT, SCRIPT, STYLE, TEMPLATE, and TITLE tags are legal children.",
         "default": false
     },
