+++ lib/presets/default.js
@@ -37,7 +37,7 @@ module.exports = {
     'id-class-ignore-regex': false,
     'img-req-alt': true,
     'img-req-src': true,
-    'head-content-model': false,
+    'head-valid-content-model': false,
     'href-style': false,
     'csslint': false,
     'label-req-for': true,
